Move cluster discovery APIs into discovery package.
