+++ cluster/src/main/java/io/atomix/cluster/discovery/MulticastDiscoveryBuilder.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.cluster;
+package io.atomix.cluster.discovery;
 
 import java.time.Duration;
