+++ cluster/src/main/java/io/atomix/cluster/discovery/MulticastDiscoveryConfig.java
@@ -13,69 +13,46 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.cluster;
+package io.atomix.cluster.discovery;
 
 import java.time.Duration;
-import java.util.Collection;
-import java.util.Collections;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
 /**
- * Bootstrap discovery configuration.
+ * Multicast discovery configuration.
  */
-public class BootstrapDiscoveryConfig extends NodeDiscoveryConfig {
-  private static final int DEFAULT_HEARTBEAT_INTERVAL = 1000;
+public class MulticastDiscoveryConfig extends NodeDiscoveryConfig {
+  private static final int DEFAULT_BROADCAST_INTERVAL = 1000;
   private static final int DEFAULT_FAILURE_TIMEOUT = 10000;
   private static final int DEFAULT_PHI_FAILURE_THRESHOLD = 10;
 
-  private Duration heartbeatInterval = Duration.ofMillis(DEFAULT_HEARTBEAT_INTERVAL);
+  private Duration broadcastInterval = Duration.ofMillis(DEFAULT_BROADCAST_INTERVAL);
   private int failureThreshold = DEFAULT_PHI_FAILURE_THRESHOLD;
   private Duration failureTimeout = Duration.ofMillis(DEFAULT_FAILURE_TIMEOUT);
-  private Collection<Node> nodes = Collections.emptySet();
 
   @Override
   public NodeDiscoveryProvider.Type getType() {
-    return BootstrapDiscoveryProvider.TYPE;
+    return MulticastDiscoveryProvider.TYPE;
   }
 
   /**
-   * Returns the configured bootstrap nodes.
+   * Returns the broadcast interval.
    *
-   * @return the configured bootstrap nodes
+   * @return the broadcast interval
    */
-  public Collection<Node> getNodes() {
-    return nodes;
+  public Duration getBroadcastInterval() {
+    return broadcastInterval;
   }
 
   /**
-   * Sets the bootstrap nodes.
+   * Sets the broadcast interval.
    *
-   * @param nodes the bootstrap nodes
-   * @return the bootstrap provider configuration
-   */
-  public BootstrapDiscoveryConfig setNodes(Collection<Node> nodes) {
-    this.nodes = nodes;
-    return this;
-  }
-
-  /**
-   * Returns the heartbeat interval.
-   *
-   * @return the heartbeat interval
-   */
-  public Duration getHeartbeatInterval() {
-    return heartbeatInterval;
-  }
-
-  /**
-   * Sets the heartbeat interval.
-   *
-   * @param heartbeatInterval the heartbeat interval
+   * @param broadcastInterval the broadcast interval
    * @return the group membership configuration
    */
-  public BootstrapDiscoveryConfig setHeartbeatInterval(Duration heartbeatInterval) {
-    this.heartbeatInterval = checkNotNull(heartbeatInterval);
+  public MulticastDiscoveryConfig setBroadcastInterval(Duration broadcastInterval) {
+    this.broadcastInterval = checkNotNull(broadcastInterval);
     return this;
   }
 
@@ -94,7 +71,7 @@ public class BootstrapDiscoveryConfig extends NodeDiscoveryConfig {
    * @param failureThreshold the failure detector threshold
    * @return the group membership configuration
    */
-  public BootstrapDiscoveryConfig setFailureThreshold(int failureThreshold) {
+  public MulticastDiscoveryConfig setFailureThreshold(int failureThreshold) {
     this.failureThreshold = failureThreshold;
     return this;
   }
@@ -114,7 +91,7 @@ public class BootstrapDiscoveryConfig extends NodeDiscoveryConfig {
    * @param failureTimeout the base failure timeout
    * @return the group membership configuration
    */
-  public BootstrapDiscoveryConfig setFailureTimeout(Duration failureTimeout) {
+  public MulticastDiscoveryConfig setFailureTimeout(Duration failureTimeout) {
     this.failureTimeout = checkNotNull(failureTimeout);
     return this;
   }
