+++ cluster/src/main/java/io/atomix/cluster/discovery/NodeDiscoveryService.java
@@ -13,12 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.cluster;
+package io.atomix.cluster.discovery;
 
-import io.atomix.utils.Managed;
+import io.atomix.cluster.Node;
+import io.atomix.utils.event.ListenerService;
+
+import java.util.Set;
 
 /**
- * Managed node discovery service.
+ * Node discovery service.
  */
-public interface ManagedNodeDiscoveryService extends NodeDiscoveryService, Managed<NodeDiscoveryService> {
+public interface NodeDiscoveryService extends ListenerService<NodeDiscoveryEvent, NodeDiscoveryEventListener> {
+
+  /**
+   * Returns the set of active nodes.
+   *
+   * @return the set of active nodes
+   */
+  Set<Node> getNodes();
+
 }
