+++ core/src/test/java/io/atomix/core/AtomixConfigTest.java
@@ -19,8 +19,8 @@ import io.atomix.cluster.ClusterConfig;
 import io.atomix.cluster.MemberConfig;
 import io.atomix.cluster.MembershipConfig;
 import io.atomix.cluster.MulticastConfig;
-import io.atomix.cluster.MulticastDiscoveryConfig;
-import io.atomix.cluster.MulticastDiscoveryProvider;
+import io.atomix.cluster.discovery.MulticastDiscoveryConfig;
+import io.atomix.cluster.discovery.MulticastDiscoveryProvider;
 import io.atomix.core.map.AtomicMapConfig;
 import io.atomix.core.profile.ConsensusProfile;
 import io.atomix.core.profile.ConsensusProfileConfig;
